# openapi.yaml
openapi: 3.0.0
info:
  title: aulaplus API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3014/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /colegio:
    get:
      tags:
        - Colegio
      summary: Listar todos los Colegios
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Colegios
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Colegio
      summary: Crear Colegio
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Colegio creado
  /colegio/{id}:
    get:
      tags:
        - Colegio
      summary: Obtener Colegio por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Colegio encontrado
        '404':
          description: Colegio no encontrado
    patch:
      tags:
        - Colegio
      summary: Actualizar Colegio
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Colegio actualizado
    delete:
      tags:
        - Colegio
      summary: Eliminar Colegio
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Colegio eliminado
  /colegio/{id}/upload-image:
    post:
      tags:
        - Colegio
      summary: Subir imagen a Colegio
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...
  /curso:
    get:
      tags:
        - Curso
      summary: Listar todos los Cursos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Cursos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Curso
      summary: Crear Curso
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Curso creado
  /curso/{id}:
    get:
      tags:
        - Curso
      summary: Obtener Curso por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Curso encontrado
        '404':
          description: Curso no encontrado
    patch:
      tags:
        - Curso
      summary: Actualizar Curso
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Curso actualizado
    delete:
      tags:
        - Curso
      summary: Eliminar Curso
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Curso eliminado
  /curso/{id}/upload-image:
    post:
      tags:
        - Curso
      summary: Subir imagen a Curso
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /evaluacion:
    get:
      tags:
        - Evaluacion
      summary: Listar todos los Evaluacions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Evaluacions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Evaluacion
      summary: Crear Evaluacion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Evaluacion creado
  /evaluacion/{id}:
    get:
      tags:
        - Evaluacion
      summary: Obtener Evaluacion por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Evaluacion encontrado
        '404':
          description: Evaluacion no encontrado
    patch:
      tags:
        - Evaluacion
      summary: Actualizar Evaluacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Evaluacion actualizado
    delete:
      tags:
        - Evaluacion
      summary: Eliminar Evaluacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Evaluacion eliminado
  /evaluacion/{id}/upload-image:
    post:
      tags:
        - Evaluacion
      summary: Subir imagen a Evaluacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida